+++ frontend/src/api/client.js
@@ -1,0 +1,15 @@
+import axios from 'axios'
+import { token } from '../store/user'
+
+const api = axios.create({
+  baseURL: import.meta.env.VITE_API_BASE_URL || '/api'
+})
+
+api.interceptors.request.use(config => {
+  if (token.value) {
+    config.headers.Authorization = `Bearer ${token.value}`
+  }
+  return config
+})
+
+export default api
