+++ frontend/src/api/user.js
@@ -1,0 +1,4 @@
+import api from './client'
+
+export const getProfile = () => api.get('/user/profile')
+export const updateProfile = data => api.put('/user/profile', data)
